requirements.txt
asgiref==3.5.0
attrs==21.4.0
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-spectacular==0.21.2
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.4.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
mysqlclient==2.1.0
pymysql==1.0.2
